// src/utils/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      diabetes_logs: {
        Row: {
          blood_glucose: number | null
          carbohydrates: number | null
          created_at: string
          exercise_duration: string | null
          exercise_name: string | null
          id: string
          insulin_basal: number | null
          insulin_bolus: number | null
          measurement_time: string | null
          profile_id: string
        }
        Insert: {
          blood_glucose?: number | null
          carbohydrates?: number | null
          created_at: string
          exercise_duration?: string | null
          exercise_name?: string | null
          id?: string
          insulin_basal?: number | null
          insulin_bolus?: number | null
          measurement_time?: string | null
          profile_id?: string
        }
        Update: {
          blood_glucose?: number | null
          carbohydrates?: number | null
          created_at?: string
          exercise_duration?: string | null
          exercise_name?: string | null
          id?: string
          insulin_basal?: number | null
          insulin_bolus?: number | null
          measurement_time?: string | null
          profile_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "diabetes_logs_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          birthday: string
          cgm: string | null
          diabetes_type: string
          first_name: string
          gender: string
          glucose_meter: string | null
          id: string
          insulin_type: string | null
          last_name: string
          year_of_diagnosis: number
        }
        Insert: {
          birthday: string
          cgm?: string | null
          diabetes_type: string
          first_name: string
          gender: string
          glucose_meter?: string | null
          id: string
          insulin_type?: string | null
          last_name: string
          year_of_diagnosis: number
        }
        Update: {
          birthday?: string
          cgm?: string | null
          diabetes_type?: string
          first_name?: string
          gender?: string
          glucose_meter?: string | null
          id?: string
          insulin_type?: string | null
          last_name?: string
          year_of_diagnosis?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never